requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
PyYAML>=6.0.1
SQLAlchemy>=2.0
langgraph>=0.2
langchain-core>=0.3
pydantic-settings>=2.2
python-multipart>=0.0.9
qdrant-client>=1.9
fastembed>=0.3
docling>=2.0

